Fixes after review 3

+++ hw2/src/main/java/ru/vitalib/otus/homework/model/Score.java
@@ -1,21 +1,21 @@
 package ru.vitalib.otus.homework.model;
 
 public class Score {
-  private final long totalQuestions;
-  private final long questionsWithCorrectAnswer;
+  private final int totalQuestions;
+  private final int questionsWithCorrectAnswer;
   private final boolean hasPass;
 
-  public Score(Long totalQuestions, long questionsWithCorrectAnswer, boolean hasPass) {
+  public Score(int totalQuestions, int questionsWithCorrectAnswer, boolean hasPass) {
     this.totalQuestions = totalQuestions;
     this.questionsWithCorrectAnswer = questionsWithCorrectAnswer;
     this.hasPass = hasPass;
   }
 
-  public long getTotalQuestions() {
+  public int getTotalQuestions() {
     return totalQuestions;
   }
 
-  public long getQuestionsWithCorrectAnswer() {
+  public int getQuestionsWithCorrectAnswer() {
     return questionsWithCorrectAnswer;
   }
